pyproject.toml
[tool.poetry]
name = "acs-cash"
version = "1.0.0"
description = "Managing cash and customers"
authors = ["kostasvit <kostasvit@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.6.0"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "3.1.0"
crispy-bootstrap4 = "2022.1"
cryptography = "40.0.1"
defusedxml = "0.7.1"
diff-match-patch = "20200713"
django = "4.2"
django-allauth = "0.54.0"
django-cors-headers = "3.14.0"
django-crispy-forms = "2.0"
django-crontab = "0.7.1"
django-encrypted-model-fields = "0.6.5"
django-environ = "0.10.0"
django-filter = "23.2"
django-import-export = "3.1.0"
django-otp = "1.2.2"
django-ranged-response = "0.2.0"
django-widget-tweaks = "1.4.12"
djangorestframework = "3.14.0"
environs = "9.5.0"
et-xmlfile = "1.1.0"
gunicorn = "20.1.0"
idna = "3.4"
markuppy = "1.14"
marshmallow = "3.19.0"
oauthlib = "3.2.2"
odfpy = "1.4.1"
openpyxl = "3.1.2"
packaging = "23.1"
pillow = "9.5.0"
psycopg2-binary = "2.9.9"
pycparser = "2.21"
pyjwt = "2.7.0"
pypng = "0.20220715.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python3-openid = "3.2.0"
pytz = "2023.3"
pyyaml = "6.0"
qrcode = "7.4.2"
requests = "2.28.2"
requests-oauthlib = "1.3.1"
six = "1.16.0"
sqlparse = "0.4.3"
tablib = "3.4.0"
typing-extensions = "4.7.1"
tzdata = "2023.3"
urllib3 = "1.26.15"
xlrd = "2.0.1"
xlwt = "1.3.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
